+++ Nostra Utils/src/source/nostrautils/core/Time.cpp
@@ -7,7 +7,7 @@ namespace NOU::NOU_CORE
 {
 	uint32 currentTimeMillis()
 	{
-		return std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count();
+		return static_cast<uint32>(std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::system_clock::now().time_since_epoch()).count());
 	}
 
 	uint64 currentTimeNanos()